# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: GIS-Agent Platform API
  description: GIS Agent 平台 — 智能方案生成系统后端 API
  version: 0.1.0
  contact:
    name: GIS Agent Team

servers:
  - url: http://localhost:8080/api
    description: 本地开发服务器

security:
  - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # ===== 通用 =====
    Error:
      type: object
      properties:
        error:
          type: string

    # ===== 认证 =====
    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 6
          maxLength: 128

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT Token
        username:
          type: string
        userId:
          type: integer
          format: int64

    # ===== LLM Provider =====
    LlmProviderCreate:
      type: object
      required: [name, providerType, endpoint]
      properties:
        name:
          type: string
          example: "GPT-4o"
        providerType:
          type: string
          enum: [openai_compatible, local]
        endpoint:
          type: string
          format: uri
          example: "https://api.openai.com/v1"
        apiKey:
          type: string
        isDefault:
          type: boolean

    LlmProviderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        providerType:
          type: string
        endpoint:
          type: string
        isDefault:
          type: boolean
        hasApiKey:
          type: boolean

    LlmTestResponse:
      type: object
      properties:
        success:
          type: boolean
        responseTimeMs:
          type: integer
        message:
          type: string

    # ===== IMA Config =====
    ImaConfigCreate:
      type: object
      required: [kbId, kbName, purpose]
      properties:
        kbId:
          type: string
          description: IMA 知识库 ID
        kbName:
          type: string
        kbType:
          type: string
          enum: [subscribed, owned]
          default: subscribed
        purpose:
          type: string
          enum: [product_doc, case_lib, industry_standard, competitor, general]
        searchWeight:
          type: number
          minimum: 0
          maximum: 1
          default: 0.5

    ImaConfigUpdate:
      type: object
      properties:
        kbName:
          type: string
        purpose:
          type: string
        searchWeight:
          type: number
          minimum: 0
          maximum: 1
        enabled:
          type: boolean

    ImaConfigResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        kbId:
          type: string
        kbName:
          type: string
        kbType:
          type: string
        purpose:
          type: string
        searchWeight:
          type: number
        enabled:
          type: boolean

    ImaTestResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string

    # ===== 项目 =====
    ProjectCreate:
      type: object
      required: [name, templateId]
      properties:
        name:
          type: string
        description:
          type: string
        templateId:
          type: string
          enum: [quick_selection, full_solution]

    ProjectResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        status:
          type: string
          enum: [DRAFT, RUNNING, COMPLETED, FAILED]
        templateId:
          type: string
        createdAt:
          type: string
          format: date-time

    PipelineStatus:
      type: object
      properties:
        pipelineRunId:
          type: integer
          format: int64
        status:
          type: string
          enum: [PENDING, RUNNING, SUCCESS, FAILED, PARTIAL]
        tools:
          type: array
          items:
            $ref: '#/components/schemas/ToolStatus'
        context:
          type: object
          description: 共享上下文数据

    ToolStatus:
      type: object
      properties:
        toolType:
          type: string
        toolOrder:
          type: integer
        status:
          type: string
          enum: [PENDING, RUNNING, SUCCESS, FAILED, SKIPPED]
        output:
          type: object
        errorMessage:
          type: string

paths:
  # ==================== 认证 ====================
  /auth/register:
    post:
      tags: [认证]
      summary: 用户注册
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: 注册成功，返回 JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: 用户名已存在或参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      tags: [认证]
      summary: 用户登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: 用户名或密码错误

  # ==================== LLM Provider ====================
  /providers:
    get:
      tags: [LLM Provider]
      summary: 获取当前用户的 LLM Provider 列表
      responses:
        '200':
          description: Provider 列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LlmProviderResponse'
    post:
      tags: [LLM Provider]
      summary: 添加 LLM Provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LlmProviderCreate'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmProviderResponse'

  /providers/{id}:
    delete:
      tags: [LLM Provider]
      summary: 删除 LLM Provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 删除成功

  /providers/{id}/test:
    post:
      tags: [LLM Provider]
      summary: 测试 Provider 连接
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 测试结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmTestResponse'

  # ==================== IMA Config ====================
  /ima/configs:
    get:
      tags: [IMA 配置]
      summary: 获取当前用户的 IMA 知识库配置列表
      responses:
        '200':
          description: 配置列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ImaConfigResponse'
    post:
      tags: [IMA 配置]
      summary: 添加 IMA 知识库配置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImaConfigCreate'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImaConfigResponse'

  /ima/configs/{id}:
    put:
      tags: [IMA 配置]
      summary: 更新 IMA 知识库配置
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImaConfigUpdate'
      responses:
        '200':
          description: 更新成功
    delete:
      tags: [IMA 配置]
      summary: 删除 IMA 知识库配置
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 删除成功

  /ima/configs/{id}/test:
    post:
      tags: [IMA 配置]
      summary: 测试知识库连接
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 测试结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImaTestResponse'

  # ==================== 项目 & 流水线 ====================
  /projects:
    post:
      tags: [项目]
      summary: 创建项目并上传需求文档
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, templateId, file]
              properties:
                name:
                  type: string
                description:
                  type: string
                templateId:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'

  /projects/{id}/run:
    post:
      tags: [流水线]
      summary: 启动流水线执行
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 流水线已启动
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineStatus'

  /projects/{id}/status:
    get:
      tags: [流水线]
      summary: 获取流水线执行状态（支持 SSE 推送）
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 流水线状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineStatus'
            text/event-stream:
              schema:
                type: string

  /projects/{id}/download/docx:
    get:
      tags: [导出]
      summary: 下载 Word 方案文档
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: .docx 文件
          content:
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary

  /projects/{id}/download/md:
    get:
      tags: [导出]
      summary: 下载 Markdown 方案文档
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: .md 文件
          content:
            text/markdown:
              schema:
                type: string